+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Breaking changes:
 
 Features:
+- [#1550](https://github.com/rails-api/active_model_serializers/pull/1550) Add
+  Rails url_helpers to `SerializationContext` for use in links. (@remear, @bf4)
 - [#1004](https://github.com/rails-api/active_model_serializers/pull/1004) JSON API errors object implementation.
   - Only implements `detail` and `source` as derived from `ActiveModel::Error`
   - Provides checklist of remaining questions and remaining parts of the spec.